Kill the --debug option.
It is confusing when there is already a ``log.level = DEBUG`` option in
the config which has nothing to do with **this** debug option.

I think we added this flag back when we first added multiprocessing
support as a mechanism for devs to debug that stuff.. but it's been in
place and working well now for a long time.  Let's kill this.