Fresh config for every test unit.
The difference in the time the tests take is negligible and it
simplifies things quite a bit. Should we start doing more in the
configuration setup such that it becomes a performance issue, I still
think we'd be better off with this approach and then only doing teardown
if the temporary directory has been modified.