Support enterprize github
This adds support for enterprize github by adding a config option
`github.host` which allows to specify the location of their github
installation. It defaults to the public github service, github.com, so
it should be backward compatible.

According to the documentation, the api endpoinds for enterprize github
are not under https://api.<hostname> but https://<hostname>/api/v3.

See https://developer.github.com/v3/enterprise/

The regular expression used for extracting the repository from the issue
was a bit too strict. This makes it more robust and  adds unit tests for
it.

The JSON returned by github when doing a search differs from what is
returned when getting issues from other api end point. In particular,
the repository url is under the key 'repository_url' instead of
'repos_url'.

This commit also adds a bunch of tests for refactored methods.

Finally, it makes nosetests more verbose: in case the test-suite hangs
in travis, making nosetests more verbose should show which test
exactly is hanging.