Put the BugwarriorData instance in the config object
Instead of each service creating their own instance of BugwarriorData, a
single instance is created and passed arround as part of the config
object.

This allows for more control for how the BugwarriorData object is
initialized in service tests. In particular, if taskwarrior isn't
initialized, instanciating a service would hang because the task command
is waiting for an input (on first run it asks if it should create a
default taskrc).