Allow limitless description and annotation lengths
This is presented as an alternative to #436.

This could lead to more-difficult-to-debug errors if people pass no
value to other configuration options, because the error will likely be
in bugwarrior failing to handle None rather than ConfigParser raising a
validation error on their config file. However, I don't see a clear
alternative. I did play around with subclassing ConfigParser to just
allow no value for integers, but the extra complication might not be
worth it.