Fix common service config. options: default_priority + add_tags
These common service configuration options were not working since
the "if 'option' in self.config:" instruction (lines 88 and 95 of
bugwarrior/services/__init__.py) was always returning False. This
occurred because these options were being fetched from a
ServiceConfig object, which requires that these options are given
in the "service.option" format (e.g. github.add_tags).

Consequently, using the "default_priority" and "add_tags" options
without the service prefix (as described in the Bugwarrior
documentation), resulted in IssueService.add_tags being set to an
empty list and IssueService.default_priority being set to the
default 'M' regardless of what was specified in the configuration.

In the proposed fix these options are now accessed directly from
the BugwarriorConfigParser object (variable main_config in the
constructor of IssueService), bypassing the ServiceConfig object.