Gmail oauth flow shouldn't attempt to parse the command line.
If any command line arguments are passed to bugwarrior, gmail will try
to parse them and fail. It's not possible pass arguments to Oauth this
way anyway (because bugwarrior would reject the unrecognised command
line arguments), so we shouldn't try to do it. (I see no reason to pass
flags to OAuth here, if needed we'd have to find another way.)