gitlab: filter repositories by regex
This introduces the service options gitlab.include_regex and
gitlab.exclude_regex, which can be used in addition to the existing
lists in gitlab.include_repos and gitlab.exclude_repos.

We have to remodel the inclusion logic a bit if we want both the regex
and the list to act as a union. Otherwise the include list would
overwrite the include regex, or vice versa.

Signed-off-by: Roland Hieber <rohieb@rohieb.name>