Fix for jira sprint parsing bug
Jira doesn't escape commas in the value field of the sprint string so we
can't just split on all commas.

In my case, I ran into a sprint with a goal that contained a bunch of
commas.

Splitting on the '=' character and then using rsplit to only grab the
last ',' is a bit safer.