Do not abort with ValueError when default_permission is blank
I use the redmine service. I have

  redmine.default_priority=

in my bugwarriorrc file, which is intended to mean "do not set an explicit priority" on tickets synced from Redmine.

I needed this change to avoid bugwarrior aborting with

  ValueError: '' is not a valid choice; choices: [None, 'H', 'M', 'L']

because the empty string needs to be mapped to None. I can't put None in the bugwarriorrc file, because it gets interpreted as a string.