BugwarriorConfigParser default allow_no_value=True
By moving the flag into the class itself we ensure we're using it in
tests.

In the case of pivotaltracker, I found that six.moves.configparser
actually behaves differently -- the real configparser throws an
exception if you try to set a non-string value.

I took the opportunity to clean up and normalize imports in the test
files I was already messing with.