Update configuration docs (#848)
* redmine: remove user_id option from docs

This option was removed from the service code a while back.

* bts: correctly document default udd_ignore_sponsor

The actual default is True while the docs suggest it's False. It
probably makes more sense to default to False but that may not be worth
changing at this point.

* bugzilla: document advanced option

All options should be in the documentation which makes the large code
comment unnecessary.

* Document project_owner_prefix when supported

* jira: document missing default values

* pagure: correctly document tag_template option

* pivotaltracker: fix annotation_comments docs

This is actually a main section option rather than service-scoped.

* pivotaltracker: fix import_blockers default docs

* pivotaltracker: fix default blocker_template

The actual default included a typo ("Resovled"). I changed it to
"State" as stated in the docs.

* taiga: document label_template option

* teamlab: correct project_name option documentation

This is an optional feature and the docs should reflect that.

* youtrack: document anonymous option

* youtrack: correct import_tags default in docs

* docs: fix typos

* docs: improve formatting of certain services

* configuration.rst: make example a valid config