config: break into package
Breaking this into a package helps us separate the different kinds of
things we were doing in this module which are not terribly related.

This also eliminates the circular dependency between
services/__init__.py and config by removing secret related functions to
a leaf module that doesn't import any other bugwarrior modules.