Validate configuration with pydantic schemas.
- Options are now type-checked.
- "Extra" options (which would have no effect) throw an error.
- Explicit service schemas make it much easier to align documentation.
(See #847 and #848 for examples of inconsistencies this uncovered.)
- My choice of pydantic was somewhat arbitrary. I've used it before and
it seems to be one of the most popular, growing, and well-maintained
validation libraries. I'd like to note that the work required to switch
to an alternative appears to be much less *after* these architectural
changes than before because they all seem to have a fairly similar api.
For example:
    - <https://www.attrs.org>
    - <https://marshmallow.readthedocs.io>
    - <https://schematics.readthedocs.io>
    - <https://github.com/keleshev/schema>