Test that example configuration is valid.
The only reason it already is valid is that I'd manually validated it
and fixed the bugs as part of #848. This will ensure we don't introduce
errors going forward.

In order to test it I moved it into a separate file which I pull in
via an include directive, which was a good opportunity to add syntax
highlighting.