gitlab: Custom query strings.
Allow users to configure custom query strings for issues, merge_requests
and todos. This allows fine-granular customization without blowing up
the configuration interface.

When a custom query is defined, issues and merge_requests are not
fetched from the repository list, but rather via the direct issue /
merge_request API. This can be magnitudes faster depending on the amount
of projects visible to the specific user and the scope of issues that
should get queried (for example querying all issues assigned to a user).