Deprecate legacy cli.
My initial thought was that there was little cost in supporting the
legacy cli indefinitely, but upon further reflection I think that would
lead to a lot of headaches on the part of maintainers and end users
alike because distributions will have varying levels of support for the
old and new cli. Some will provide both, some will not notice the new
cli and continue only providing the old cli, and some will remove the
old cli in favor of the new. This divergence in support will cause a lot
of confusion when folks can't find the command they're looking for after
installing bugwarrior with their package manager.