github: fetch assigned issues from all repos
Fix #375.

I'm guessing the semantics of this api endpoint have changed since we started
using it as a lot of these different repository management schemes didn't yet
exist.

Per the documentation for the relevant "/user/issues" api assigned issues will
only be pulled if you are the owner or a member.
<https://docs.github.com/en/rest/reference/issues#list-user-account-issues-assigned-to-the-authenticated-user>

The "/issues" api broadens this to include all "visible" issues, which I would
assume includes all public repositories. I believe the documentation is
unclear on this simply because they primarily have the enterprise user
in mind, for which public repos are irrelevant.
<https://docs.github.com/en/rest/reference/issues#list-user-account-issues-assigned-to-the-authenticated-user>