Decrement currently_running one time. Fix #897.
Unfortunately this was pointed out by @fmauch but it appears that I
accidentally reverted it when rebasing #825.

> The currently_running -= 1 will also be reached once yield is called,
> as the generator simply continues directly after the yield statement
> (as far as I understood generators).

https://github.com/ralphbean/bugwarrior/pull/825#discussion_r659296257

This explains the situation of #897:

> We hit the line log.exception("Worker for [%s] failed: %s" % (target, e))
> but then somehow miss the line log.exception(f"Aborted {target} due to
> critical error."

https://github.com/ralphbean/bugwarrior/issues/897#issuecomment-1096838119

Because the decrement was occurring twice for a previous service, the
`while currently_running > 0` loop was ending before getting  to the
`SERVICE_FINISHED_ERROR` in the queue.

Also, improve logging consistency in this section.