gitlab: Remove support for disabled features.
By my understanding all supported releases of gitlab should return an
empty list even if these features aren't supported. Regardless it's not
evident that there is any way to distinguish between a failure due to
lack of gitlab server support and any other kind of network failure. If
this is actually still needed we either need to find a way to
distinguish such exceptions or add an option to toggle these features
off.