config/schema: Strip prefixes before validation
Adding a new configuration file format seems like an appropriate time to
drop the "prefix" requirement.

Along with stripping prefixes, we're also converting the ConfigParser
instance to a dictionary before passing it into validation, which will
be necessary so that future configuration parsers can also pass this
universal format. This means that almost everywhere in the test suite we
were building ConfigParser's can now just use a dictionary.

Since the schema is now unaware of configparser features we remove the
requirement of a DEFAULT section, which is included in any ConfigParser
instance but will not be included in the dictionaries provided by other
configuration language parsers.