config/schema: Strip prefixes before validation
Adding a new configuration file format seems like an appropriate time to
drop the "prefix" requirement. While we want to maintain backwards
compatibility, maintaining "prefix awareness" within the validation
which is only relevant during INI validation would require even more
intricate metaprogramming and I don't think it's worth it. It's
questionable whether the existing amount of error message enhancement is
worth the future maintenance burden, so I think deleting some of it is
worth the cost of somewhat degrading the error messages. They're still
better than they were in the last release.

As an additional side-effect though, bugwarrior will no longer care if
prefixes are correct or if they're even present for INI files. If I
recall correctly, in previous releases these were be silently ignored,
but now they'll "just work". I'm not sure it really matters...

Along with stripping prefixes, we're also converting the ConfigParser
instance to a dictionary before passing it into validation, which will
be necessary so that future configuration parsers can also pass this
universal format. This means that almost everywhere in the test suite we
were building ConfigParser's can now just use a dictionary.

Since the schema is now unaware of configparser features we remove the
requirement of a DEFAULT section, which is included in any ConfigParser
instance but will not be included in the dictionaries provided by other
configuration language parsers.