configuration: toml support. Resolve #873
Add tomli as a backport of the standard library's new tomllib.

By the way, it's not just a convention to use the `.toml` extension,
it's actually part of the v1.0.0 spec:

> TOML files should use the extension `.toml`.

So we don't need to do any additional logic to determine whether a file
is ini or toml -- if it doesn't have the right extension it literally
isn't toml!