tests/test_load: test config path combinations
Add pytest-subtests plugin to test dependencies so that pytest will show
the parameters when a subTest fails.

Replace methods that check against specific combinations.

It would be a lot simpler to use pytest.mark.parametrize, which would
eliminate the need for pytest-subtests as well as the manual
setUp/tearDown, but that feature is unavailable for classes inheriting
from TestCase. Converting every class that inherits from ConfigTest to
pytest assertions would be too much extra noise to add to this branch,
which already has a large diff. I'm also not sure it's worth the effort
since it seems like it will take a couple more decades for python to
standardize a testing API. ("setup.py is deprecated; use pyproject.toml
and pick one of the 5 incompatible formats; use pytest as a test runner;
but unittest is in the standard library with a partially incompatible
API...")