config: Deprecate ini in favor of toml.
If we're going to degrade the validation and drop the documentation, the
best favor we can do users is to deprecate ini configuration altogether
as it is not receiving first-class support.

I took this opportunity to remove the ini example configuration file
from the table of contents and replace it with an extremely brief
example at the beginning of the configuration documentation. In my
opinion this is a long overdue change because a 200+ line configuration
file is a poor excuse for documentation. If there's anything about the
configuration format it covers that the documentation otherwise does
not, I'm sure we can express it in less than 200 lines.