config/load: Use ini2toml to simplify ini parsing.
This may be taking it a bit too far, but it does have its advantages.
The error-prone parsing and translation of ini files is now centralized
in one place which should make it easier to weed out the bugs and has
the additonal benefit of removing ini-specific code from the generic
configuration loading module. It also brings back the prefix validation
errors which had to be removed from the schema itself.