pull: fix "--interactive" flag
The string value was truthy even when it was "False". Whoops.

It would be nice to have a regression test for this but it's not obvious
how to do that. However, I do plan to start mutating these values on the
configuration before validation in an upcoming pull request, which would
have caught this because it's the wrong type.