IssueService.get_templates -> ServiceConfig.compute_templates
Somewhat following the logic of #976, computed fields should be added to
the configuration object when practical.

This is motivated by slimming down the base service classes for #775 by
removing the idiosyncratic `get_template()` method, which is the only
thing passed to `Issue.origin` that wasn't already coming directly from
a configuration field.