gitlab: Begin migrating owned to default to True
Per discussion in #986, this has a few benefits:

- Our example configuration will actually be valid since it uses
gitlab.com as a host but does not define a filter.
- More similar default behavior to other services, such as github.
- Hopefully less confusion about the filtering requirements to
avoid rate-banning and fewer bug reports caused by rate bans.

We start off by implementing the first step, which is to raise a
warning if 'owned' is undefined. The remainder of the migration plan is
in the validator docstring.