gitbug: Correctly expand user home directory
Previously, providing '~' as part of the path to a gitbug remote would
not correctly expand to the user's home directory and error out with
directory not found.

This commit fixes the behavior to correctly expand the path instead,
aligning behavior with that of the taskrc configuration option and other
services' path expansion.