service.Issue: Remove "mapping type" methods
https://docs.python.org/3/library/stdtypes.html#mapping-types-dict

We're no longer relying on implicit `dict()` conversion and are rather
using the `get_taskwarrior_record()` method directly.

I assume the original motivation for this was convenience of use in
third-party applications which used bugwarrior as a library. As far as I
know no such applications are still in use.

In preparation for documenting and stabilizing the base class interfaces
(#791) I'd like to cut them back to a minimum. Maybe the mapping type
could be added back some day if there were demand for it, but it would
then need to conform to a standard of testing a documentation I plan to
implement.