Break TaskConstructor out of base Issue class.
In preparation for stabilizing the base classes (#791), I'm moving these
methods which child classes never need to call to a separate class in
the collect module.

We *could* leave them where they are and just make them private methods,
but at this point I think slimming down the base classes as much as
possible is our best bet to being disciplined about their stabilization.
This not only removes lines of code but I believe the entire class of
would-be private methods.

I also removed the __str__ and __repr__ methods, which depended on the
get_taskwarrior_record method. These used to be used in logging messages
in db.synchronize but since #1037 db.synchronize is receiving dict's and
not Issue's. (The usages in that method are presently mixed between
logging `issue` and logging `issue['description']`. The logs could be
improved by switching the remainder to the latter.)