Remove db.MARKUP
Originally, this description prefix was used by bugwarrior to determine
which tasks it manages and thus is responsible for closing. This is now
done entirely through each service's UNIQUE_KEY attribute.

It might be best to remove the prefix entirely from the description but
I'll leave that for another day. Maybe it's helpful to know at a glance
which tasks were added by bugwarrior?