use pydantic v2 backwards compat mode
This should allow using the existing pydantic v1 features while
migration to v2 is not done. This also raises the minimum requirement to
pydantic >= 2, as the v1 compat modules are not available in pydantic
v1. As discussed in the PR [0], this new version requirement was
accepted due to widespread adoption of pydantic v2 in major
distributions.

Issue: https://github.com/GothenburgBitFactory/bugwarrior/issues/998

[0] https://github.com/GothenburgBitFactory/bugwarrior/pull/1068#pullrequestreview-2310309470