Remove issue filtering from base Service class
We here remove the abstract `get_owner` method and the concrete
`include` method from the base class and copy the `include` method to
the handful of services which make use of it. The `include` method was
only being used in 6 of 26 services, so I think the duplication is
acceptable in exchange for the benefits:

- This eliminates the last instance of bidirectional communication
  between the base classes and the service implementations, which in my
  opinion is an architectural problem.
- Since most of the services don't use this filtering mechanism, it is
  annoying to have to define `get_owner`.
- This filtering mechanism is somewhat legacy from the days when API's
  were generally far less capable. I don't think we should encourage it
  going forward because it is better to implement filtering through the
  API query rather than pulling all data and iterating over it locally,
  and this is a reasonable expectation from third party API's these
  days.

One might suspect that eliminating these NotImplementedError's would
eliminate the warnings from services that do not support
only_if_assigned, but in fact these would only get raised if the service
itself called `self.include`, which they do not. So a user trying to use
these options already would find them silently ignored, which isn't
great but is a problem for another day.