services reference common configuration options
- Add a test that all services reference the common configuration
  options (at least vicariously). The exception is add_tags, which is done
  during the issue aggregation phase.
- Add an UnsupportedOption type to the configuration schema which throws
  a validation error if the given value is changed from it's default.
- Set the unimplemented options to this UnsupportedOption value.

One of the effects of no longer requiring the get_owner method in #1070
which I wasn't pleased with is that there is no longer an impetus for
the service developer to implement these common configuration options.
However, I also learned in that PR that making imperfect but nonetheless
useful tests by regexing the code can cheaply make some loose guarantees
that software follows some basic guidelines.

So this should restore the virtue that unimplemented features need to be
noticed and suggests an even more explicit statement that they are
unsupported. Furthermore, this gives users a validation error when they
configure one of these unsupported features, rather than failing
silently as the #1070 commit notes was always the case:

> One might suspect that eliminating these NotImplementedError's would eliminate the warnings from services that do not support only_if_assigned, but in fact these would only get raised if the service itself called self.include, which they do not. So a user trying to use these options already would find them silently ignored, which isn't great but is a problem for another day.

Today is that day!