Fix --flavor feature. Resolve #1166.
This regression was probably introduced when adding toml support. The
main issue is that the code is "indecisive" about whether flavors should
be stored as top-level fields (`"flavor.myflavor"`) or a nested dictionary
field (`"flavor": { "myflavor": ... }`). While unquoted sections are
parsed into the nested structure in toml, top-level sections seem to be
a bit easier to work with, so we now ensure that the data structure is
flattened into top-level flavor fields.

Additionally:
- ini2toml now does not result in quoted flavor sections. This makes no
  functional difference but it does align with the docs and IMO is a bit
  more attractive.
- `interactive` is now optional because otherwise we'd have to set it
  for every flavor including "inactive" ones.
- Flavor configuration is now tested in several ways whereas it was not
  exercised in the test suite at all before.